Add files via upload
This code models the light propagation in a degenerate type-I nanosecond OPO, based on the results published in "Degenerate type I nanosecond optical parametric oscillators", Smith et al. (2003). Essentially, this code solves the well-known coupled equations in a nonlinear crystal in order to obtain both temporal and spectral outputs. In this case, there are only two coupled-equations due to degeneracity.